requirements.txt
# Core dependencies
langchain==0.3.7
langchain-community==0.3.5
langchain-core==0.3.15

# Vector store
qdrant-client==1.11.3

# PDF processing
PyMuPDF==1.24.13
Pillow==10.4.0

# Ollama integration
ollama==0.3.3

# Utilities
python-dotenv==1.0.1
requests==2.32.3
tqdm==4.66.5

# Caching
diskcache==5.6.3
groq>=0.5.0
langchain-groq>=0.1.0
